6242262 ON should be buildable with gcc 6268345 fmd is insufficiently careful with alignment 6271070 gcc and cmd/fm don't get along 6272173 cpumem diag module writes to constant memory, crashing fmd 6273907 gcc and cmd/syseventd don't get along 6308057 gcc and sgs/dis don't get along on SPARC 6359863 the C preprocessor is still feeling abused by token pasting 6359868 gcc and cmd/stmsboot still don't get along 6359878 gcc and lib/smartcard don't get along on sparc 6361816 gcc and perl don't get along 6370832 kernel should agree on C99 mode for studio and gcc 6372728 gcc and sata don't get along 6390345 cw inserts garbage when used as a preprocessor 6397680 gcc and smbios are not getting along 6399113 meta_namespace.c uses worrisome construct and is not gcc friendly 6403999 cw is ugly and hairy and needs a man

+++ usr/src/cmd/fm/modules/common/eversholt/alloc.c
@@ -2,9 +2,8 @@
  * CDDL HEADER START
  *
  * The contents of this file are subject to the terms of the
- * Common Development and Distribution License, Version 1.0 only
- * (the "License").  You may not use this file except in compliance
- * with the License.
+ * Common Development and Distribution License (the "License").
+ * You may not use this file except in compliance with the License.
  *
  * You can obtain a copy of the license at usr/src/OPENSOLARIS.LICENSE
  * or http://www.opensolaris.org/os/licensing.
@@ -19,8 +18,9 @@
  *
  * CDDL HEADER END
  */
+
 /*
- * Copyright 2004 Sun Microsystems, Inc.  All rights reserved.
+ * Copyright 2006 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  *
  * alloc.c -- memory allocation wrapper functions, for eft.so FMD module
@@ -40,8 +40,8 @@
 
 extern fmd_hdl_t *Hdl;		/* handle from eft.c */
 
-/* room to store size */
-#define	HDRSIZ	sizeof (size_t)
+/* room to store size, possibly more to maintain alignment for long longs */
+#define	HDRSIZ	sizeof (long long)
 
 static struct stats *Malloctotal;
 static struct stats *Freetotal;
